moved app from __init__ to app module, as it was imported in tests, lead to the tests importing onnxruntime and fail over missing openMP
